# .api/apis/chainbase/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "🛠️ Basic API",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "https://api.chainbase.online",
      "description": "production server"
    }
  ],
  "paths": {
...
    "/v1/account/txs": {
      "get": {
        "tags": ["Transaction"],
        "summary": "Get transactions by account",
        "description": "Returns the transactions from the address.",
        "operationId": "GetAccountTxs",
        "parameters": [
          {
            "description": "chain network id, \n    see [list of details](https://docs.chainbase.com/reference/supported-chains)\n    ",
            "required": true,
            "schema": {
              "title": "Chain Id",
              "type": "integer",
              "description": "chain network id, \n    see [list of details](https://docs.chainbase.com/reference/supported-chains)\n    "
            },
            "example": "1",
            "name": "chain_id",
            "in": "query"
          },
          {
            "description": "A hex string referencing a wallet address",
            "required": true,
            "schema": {
              "title": "Address",
              "pattern": "^0[xX]{1}[a-fA-F0-9]{40}$",
              "type": "string",
              "description": "A hex string referencing a wallet address"
            },
            "example": "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045",
            "name": "address",
            "in": "query"
          },
          {
            "description": "Inclusive from block number (hex string or int)",
            "required": false,
            "schema": {
              "title": "From Block",
              "type": "string",
              "description": "Inclusive from block number (hex string or int)"
            },
            "name": "from_block",
            "in": "query"
          },
          {
            "description": "Inclusive to block number (hex string, int, or 'latest')",
            "required": false,
            "schema": {
              "title": "To Block",
              "type": "string",
              "description": "Inclusive to block number (hex string, int, or 'latest')"
            },
            "example": "latest",
            "name": "to_block",
            "in": "query"
          },
          {
            "description": "Inclusive from block number (hex string or int)",
            "required": false,
            "schema": {
              "title": "From Timestamp",
              "type": "integer",
              "description": "Inclusive from block number (hex string or int)"
            },
            "name": "from_timestamp",
            "in": "query"
          },
          {
            "description": "Inclusive end timestamp",
            "required": false,
            "schema": {
              "title": "End Timestamp",
              "type": "integer",
              "description": "Inclusive end timestamp"
            },
            "name": "end_timestamp",
            "in": "query"
          },
          {
            "description": "The page offset",
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1,
              "type": "integer",
              "description": "The page offset",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "description": "The desired page size limit. Less or equal than 100",
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "description": "The desired page size limit. Less or equal than 100",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            },
            "example": "demo",
            "name": "x-api-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ResponseGetAccountTxs",
                  "required": ["data"],
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "integer",
                      "default": 0
                    },
                    "message": {
                      "title": "Message",
                      "type": "string",
                      "default": "success"
                    },
                    "data": {
                      "title": "Data",
                      "type": "array",
                      "items": {
                        "title": "TxDetail",
                        "required": [
                          "type",
                          "status",
                          "block_number",
                          "block_timestamp",
                          "transaction_hash",
                          "transaction_index",
                          "from_address",
                          "to_address",
                          "value",
                          "input",
                          "nonce",
                          "contract_address",
                          "gas",
                          "gas_price",
                          "gas_used",
                          "effective_gas_price",
                          "cumulative_gas_used",
                          "max_fee_per_gas",
                          "max_priority_fee_per_gas",
                          "tx_fee",
                          "saving_fee",
                          "burnt_fee"
                        ],
                        "type": "object",
                        "properties": {
                          "type": {
                            "title": "Type",
                            "type": "integer"
                          },
                          "status": {
                            "title": "Status",
                            "type": "integer"
                          },
                          "block_number": {
                            "title": "Block Number",
                            "type": "integer"
                          },
                          "block_timestamp": {
                            "title": "Block Timestamp",
                            "type": "string"
                          },
                          "transaction_hash": {
                            "title": "Transaction Hash",
                            "type": "string"
                          },
                          "transaction_index": {
                            "title": "Transaction Index",
                            "type": "integer"
                          },
                          "from_address": {
                            "title": "From Address",
                            "type": "string"
                          },
                          "to_address": {
                            "title": "To Address",
                            "type": "string"
                          },
                          "value": {
                            "title": "Value",
                            "type": "string"
                          },
                          "input": {
                            "title": "Input",
                            "type": "string"
                          },
                          "nonce": {
                            "title": "Nonce",
                            "type": "string"
                          },
                          "contract_address": {
                            "title": "Contract Address",
                            "type": "string"
                          },
                          "gas": {
                            "title": "Gas",
                            "type": "integer"
                          },
                          "gas_price": {
                            "title": "Gas Price",
                            "type": "integer"
                          },
                          "gas_used": {
                            "title": "Gas Used",
                            "type": "integer"
                          },
                          "effective_gas_price": {
                            "title": "Effective Gas Price",
                            "type": "integer"
                          },
                          "cumulative_gas_used": {
                            "title": "Cumulative Gas Used",
                            "type": "integer"
                          },
                          "max_fee_per_gas": {
                            "title": "Max Fee Per Gas",
                            "type": "integer"
                          },
                          "max_priority_fee_per_gas": {
                            "title": "Max Priority Fee Per Gas",
                            "type": "integer"
                          },
                          "tx_fee": {
                            "title": "Tx Fee",
                            "type": "integer"
                          },
                          "saving_fee": {
                            "title": "Saving Fee",
                            "type": "integer"
                          },
                          "burnt_fee": {
                            "title": "Burnt Fee",
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "next_page": {
                      "title": "Next Page",
                      "type": "integer"
                    },
                    "count": {
                      "title": "Count",
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ErrorMessage",
                  "required": ["code", "message"],
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "integer"
                    },
                    "message": {
                      "title": "Message",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ErrorMessage",
                  "required": ["code", "message"],
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "integer"
                    },
                    "message": {
                      "title": "Message",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ErrorMessage",
                  "required": ["code", "message"],
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "integer"
                    },
                    "message": {
                      "title": "Message",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ErrorMessage",
                  "required": ["code", "message"],
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "integer"
                    },
                    "message": {
                      "title": "Message",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ErrorMessage",
                  "required": ["code", "message"],
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "integer"
                    },
                    "message": {
                      "title": "Message",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ErrorMessage",
                  "required": ["code", "message"],
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "integer"
                    },
                    "message": {
                      "title": "Message",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}